requirements.txt
protobuf==4.25.1
streamlit==1.28.2
streamlit_chat==0.1.1
google-ai-generativelanguage==0.3.3
google-api-core==2.14.0
google-auth==2.23.4
google-generativeai==0.2.2
googleapis-common-protos==1.61.0
torch
transformers
